Ros msgs working

+++ Gantry/main/DC_Motor.ino
@@ -3,10 +3,10 @@
 //Speeds
 extern volatile int speed_X = zeroSpeed;
 extern volatile int speed_Y = zeroSpeed;
-int Right = 20;
-int Left = 180;
-int Forward = 200;
-int Backward = 0;
+double Right = 20;
+double Left = 180;
+double Forward = 200;
+double Backward = 0;
 
 //State
 int X_count = 0;
